+++ R/methods-crop.R
@@ -44,7 +44,7 @@
 #' e <- ext(img1)
 #' e <- e - 1000
 #' 
-#' img2 <- crop(img1, e)s
+#' img2 <- crop(img1, e)
 #' plot(img2)
 NULL